cleaned up upload
